recommendation_service/requirements.txt
flask
pymongo
google-generativeai
python-dotenv
